Move delete wallet function to a reusable helper
