+++ src/lib/wallets/index.ts
@@ -99,6 +99,36 @@ export const createWallet = async () => {
   return { id }
 }
 
+export const deleteWallet = async (id: Wallet['id']) => {
+  await db.wallets.delete(id)
+
+  const connections = connections$.value
+  const connectionIndex = connections.findIndex(({ walletId }) => walletId === id)
+
+  if (connectionIndex !== -1) {
+    const connection = connections[connectionIndex]
+
+    // disconnect
+    connection.disconnect && (await connection.disconnect())
+
+    // remove from connections
+    connections.splice(connectionIndex, 1)
+
+    // update connections
+    connections$.next(connections)
+  }
+
+  await Promise.all(
+    db.tables.map(async table => {
+      try {
+        await table.where('walletId').equals(id).delete()
+      } catch (error) {
+        // wallet id is not indexed, all good to swallow error here
+      }
+    })
+  )
+}
+
 export const listenForConnectionErrors = (connection: Connection) => {
   connection.errors$.pipe(takeUntil(connection.destroy$)).subscribe(errors$)
 }
