+++ src/routes/auto-connect.ts
@@ -1,8 +1,9 @@
 import type { CoreLnConfiguration, Wallet, WalletConfiguration } from '$lib/@types/wallets.js'
 import { createRandomHex, encryptWithAES } from '$lib/crypto.js'
-import { db, deleteAll } from '$lib/db/index.js'
+import { db } from '$lib/db/index.js'
 import { larpMode$, wallets$ } from '$lib/streams.js'
 import { nowSeconds } from '$lib/utils.js'
+import { deleteWallet } from '$lib/wallets'
 
 export const autoConnectWallet = async (options: {
   configuration: WalletConfiguration
@@ -39,8 +40,9 @@ export const autoConnectWallet = async (options: {
 
     // Only one wallet can be connected in larp mode
     if (larpMode$.value && wallets$.value.length) {
-      await deleteAll()
-      await db.open()
+      for (const { id } of wallets$.value) {
+        await deleteWallet(id)
+      }
     }
     await db.wallets.add(wallet)
